+++ firefoxThemes/modules/utils/themes/background.js
@@ -1,0 +1,6 @@
+/*Check if current theme has a secondary background*/
+function hasSecondaryBG(currentTheme) {
+  return currentTheme && !!currentTheme.backgrounds.secondary;
+}
+
+export {hasSecondaryBG};